Add ask command and rendering of workspaces
The biggest fundamental change is that :hypertext/answer to refers to a
pointer. Later I might have to introduce several pointer types.

The code is becoming repetitive in places. I'll accept this for now and
refactor when the need arises. I feel that all my design energy is going
into the model and little is left to make the code elegant.