jursey.core: Add orientation defs
The file is becoming big and therefore difficult to navigate. For now I
don't want to split it, because that would also be harder to work with.

The "Structure" pane in IntelliJ helps navigating between definitions.
However, in Clojure this is only a long, non-nested list. So in the code
I add ;;;; heading comments. But IntelliJ doesn't pick them up and I
couldn't quickly find a plugin that does it. Add peculiar defs as a
workaround.