schema: Make some attributes components
When I started writing Jursey I thought that hypertexts would be shared
between workspaces. Thus questions and answers could not be part of one
workspace. It was a relict of that time that the attributes weren't
component attributes.

Now I noticed that it's useful for retracting workspaces (which I don't
do after all – cf. a comment in the next commit) to have all component
attributes. So change the schema in that way.