schema: Add :qa/ws
Before a sub-question is answered, its :qa/answer refers to the
workspace where the answer will be given. After it is answered, this
reference goes away. By looking at the history DB, one could still find
the workspace where the answer was given, but that is tedious. Retain
the workspace in :qa/ws in order to avoid the roundabout way.