schema: Add :reflect/reachable
This will become necessary later when I start passing pointers to copies
of :reflect/*s around, because the time a copy is created determines
what is reachable from there. So we need to store the time somewhere.
This is in :reflect/reachable.

Already now :reflect/reachable will reduce the complexity of finding out
what is reachable.