Complete implementation of asking with reflection
Also add many tests to scenarios.repl. So far this herd of test blobs is
useful. I'm worried that it will be unmaintainable and suddenly collapse
into uselessness. But then I can write more maintainable tests if there
is time.

Another drawback is, of course, that these are just examples that came
to my mind. So they don't highlight corner cases that didn't come to my
mind. And the number of corner cases has grown a lot, as you can guess
from the growth in the number of test blobs.