unlock-by-pmap: Remove last argument
Since I call act-txreq in `unlock` and therefore outside unlock-by-pmap,
the last argument is obsolete. There's a slight kink now in that the
unlocks made by get-root-qa don't have :tx/act entries anymore. However,
the root question pseudo-workspace is not supposed to be reflected, so
this shouldn't be a problem.