Fix a defect with versions
I had assumed that versions of a workspace are synchronous with actions
in that workspace. This is wrong. Example: User sees version 1 of
workspace W and unlocks an ungiven answer. Nobody will see W until the
reply to the answer is given. So version 2 can't be the state after the
unlock. It has to be the state after the reply. However, the reply is
issued in another workspace.

Change the functions retrieving versions and actions to take this into
account.

Action              Resulting version
                    Before   Now
:ask in parent      0        0
:ask in current     1        1
:unlock in current  2        1
:reply in child     2        2