Remove pseudo-workspace parent entry
A root question is the sub-question of a pseudo-workspace (ie. a
workspace without a question). A pseudo-workspace is the parent of a
root workspace (ie. a workspace whose question is a root question).

Pseudo-workspace: {"sq" {0 {"q" "<root question>"}}}
Root workspace: {"q" "<root question>"}

So far the reflection structure of a root workspace had a "parent"
entry. However, as this parent isn't a real workspace, it shouldn't be
reflectable. I knew this before, but I didn't take the time to remove
the "parent" entry from root workspaces. Do it now.