Add tutorial and jursey.repl-ui
The tutorial can hopefully teach people how to use Jursey. The REPL UI
is a more concise way to use the single-user runner for playing around.
It will be helpful for casual users who don't have the dev tools that I
use.

While writing the tutorial and running the commands in a plain Leiningen
nREPL client, I noticed that it doesn't have pretty-printing. So add a
Leiningen plugin for pretty-printing.