README: Add acknowledgements
My attitude to this project was that it's like a contracting job. People
don't usually write acknowledgements in contracting job documentation.
But usually contracting job documentation is not public and people know
whose contributions went into it. This one is public and the public
doesn't know anything about it, so I shouldn't make it look like I was
the only contributor.